fix: change 'cost' type from double to int. 'cost' may be int

+++ src/protocols/RoutingProtocolManagementInterface.java
@@ -1,5 +1,5 @@
 public interface RoutingProtocolManagementInterface {
 	boolean getDistanceTable(short nodeId);
 	boolean getLinkCost(short nodeIdA, short nodeIdB);
-	boolean setLinkCost(short nodeIdA, short nodeIdB, double cost);
+	boolean setLinkCost(short nodeIdA, short nodeIdB, int cost);
 }
